requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-taggit==0.22.1
django-widget-tweaks==1.4.1
djangorestframework==3.6.4
idna==2.6
kombu==4.1.0
Markdown==2.6.9
oauthlib==2.0.3
olefile==0.44
Pillow==4.2.1
PyJWT==1.5.3
python-decouple==3.1
python-openid==2.2.5
python-social-auth==0.3.6
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
vine==1.1.4
